backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]
pydantic==2.11.3
python-dotenv==1.1.0
httpx==0.28.1
google-genai==1.13.0
openai==1.77.0